Passed Unique Hash User on Model Creation.

+++ server/php/api/destroyDataset.php
@@ -1,6 +1,6 @@
 <?php
     require_once "../dbconnect.php";
-    // require_once "../global_functions.php";
+    require_once "../global_functions.php";
 
     $method = $_SERVER['REQUEST_METHOD'];
 
@@ -12,18 +12,20 @@
         exit;
     }
 
-    // if(!isset($input['token'])) {
-    //     header("HTTP/1.1 400 Bad Request");
-    //     print json_encode(['errormesg'=>"Token is not set."]);
-    //     exit;
-    // }
+    // Token Validation.
+    if(!isset($input['token'])) {
+        header("HTTP/1.1 400 Bad Request");
+        print json_encode(['errormesg'=>"Token is not set."]);
+        exit;
+    }
 
-    // if(!token_exists($input['token'])) {
-    //     header("HTTP/1.1 400 Bad Request");
-    //     print json_encode(['errormesg'=>"Token doesn't exist."]);
-    //     exit;
-    // }
+    if(!token_exists($input['token'])) {
+        header("HTTP/1.1 400 Bad Request");
+        print json_encode(['errormesg'=>"Token doesn't exist."]);
+        exit;
+    }
 
+    // Folder Validation.
     if(!isset($input['folder'])) {
         header("HTTP/1.1 400 Bad Request");
         print json_encode(['errormesg'=>"Please select a folder type."]);
@@ -38,6 +40,7 @@
         exit;
     }
 
+    // Dataset Validation.
     if(!isset($input['file'])) {
         header("HTTP/1.1 400 Bad Request");
         print json_encode(['errormesg'=>"Please select a dataset to delete."]);
@@ -45,53 +48,52 @@
     }
 
     $file = $input['file'];
-
     $delete = false;
 
     if($folder == "public") {
-        // $query = 'select public_permission from users where token=?';
-        // $st = $mysqli->prepare($query);
-        // $st->bind_param('s',$input['token']);
-        // $st->execute();
-        // $res = $st->get_result();
-        // $public_permission = $res->fetch_assoc()['public_permission']; 
-        
 
-        // if($public_permission == 0) {
-        //     header("HTTP/1.1 403 Forbidden");
-        //     print json_encode(['errormesg'=>"You don't have the permission to delete public data."]);
-        //     exit;
-        // }
+        $query = 'select public_permission from users where token=?';
+        $st = $mysqli->prepare($query);
+        $st->bind_param('s', $input['token']);
+        $st->execute();
+        $res = $st->get_result();
+        $public_permission = $res->fetch_assoc()['public_permission']; 
+        
+        if($public_permission == 0) {
+            header("HTTP/1.1 403 Forbidden");
+            print json_encode(['errormesg'=>"You aren't allowed to delete public datasets."]);
+            exit;
+        }
         
         $file_path = "../../py/public/datasets/$file";
     
         if(!file_exists($file_path)) {
             header("HTTP/1.1 400 Bad Request");
-            print json_encode(['errormesg'=>"File doesn't exist."]);
+            print json_encode(['errormesg'=>"Dataset doesn't exist."]);
             exit;
         }
         
         $delete = unlink($file_path);
-
     }
     else {
-        // $email = user_mail($input['token']);
 
-        // $hash_user = md5($email);
-        // $file_path = "../../py/users/$hash_user/datasets/$file";
+        $email = user_mail($input['token']);
+        $hash_user = md5($email);
 
-        // if(!file_exists($file_path)){
-        //     header("HTTP/1.1 400 Bad Request");
-        //     print json_encode(['errormesg'=>"File doesn't exist."]);
-        //     exit;
-        // }
+        $file_path = "../../py/users/$hash_user/datasets/$file";
+
+        if(!file_exists($file_path)) {
+            header("HTTP/1.1 400 Bad Request");
+            print json_encode(['errormesg'=>"Dataset doesn't exist."]);
+            exit;
+        }
         
-        // $delete = unlink($file_path);
+        $delete = unlink($file_path);
     }
     
     if(!$delete) {
         header("HTTP/1.1 400 Bad Request");
-        print json_encode(['errormesg'=>"Unable to delete this file."]);
+        print json_encode(['errormesg'=>"Unable to delete this dataset."]);
         exit;
     }
